backend/pyproject.toml
[project]
name = "gobot-backend"
version = "0.1.0"
description = "FastAPI scraping service for GoBot"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "pydantic-settings>=2.7.0,<3.0.0",
    "playwright>=1.54.0,<2.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
]

[project.scripts]
gobot-api = "app.main:main"

[dependency-groups]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
